backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Prisma ORM
prisma==0.11.0

# Database
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Pydantic for validation
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Supabase Storage SDK
supabase==2.3.0
# Note: httpx is included as a dependency of supabase (requires <0.25.0)

# HTTP client dependencies (fix compatibility issues)
httpx==0.24.1
httpcore==0.18.0
h11==0.14.0

# Additional utilities
python-dotenv==1.0.0

# Google Gemini AI
# Option 1: Newer API (recommended)
google-genai>=0.1.0
# Option 2: Legacy API (fallback)
google-generativeai>=0.3.2
